// src/entities/Article/model/consts/articleConsts.ts
export enum ArticleSortField {
    VIEWS = "views",
    TITLE = "title",
    CREATED = "createdAt",
}

export enum ArticleBlockType {
    TEXT = "TEXT",
    CODE = "CODE",
    IMAGE = "IMAGE",
}

export enum ArticleType {
    ALL = "ALL",
    IT = "IT",
    SCIENCE = "SCIENCE",
    ECONOMICS = "ECONOMICS",
}

export enum ArticleView {
    BIG = "BIG",
    SMALL = "SMALL",
}